Move html stuff from Base to HtmlBaseElement
And move magic stuff with class/id into Element. Need to check later if
we really need id.
